requirements.txt
fastapi
uvicorn
pydantic
requests
Pillow
python-dotenv
opencv-python
easyocr
pytesseract
nest-asyncio
PyMuPDF
python-multipart
torch
langchain
langchain_chroma
langchain_core
langchain_huggingface
langchain_openai
sentence-transformers
chromadb
llama-index
langchain-community
llama_parse
layoutparser
scikit-learn
textstat
inflect
nltk
